Add phoneme support and phonics breakdown
