+++ src/GameScreen.tsx
@@ -26,6 +26,7 @@ import HintPanel from "./components/HintPanel";
 import AvatarSelector from "./components/AvatarSelector";
 import { AudioSettings } from "./components/AudioSettings";
 import { useAudio } from "./AudioContext";
+import PhonicsBreakdown from "./components/PhonicsBreakdown";
 
 const musicStyles = ['Funk', 'Country', 'Deep Bass', 'Rock', 'Jazz', 'Classical'];
 
@@ -645,6 +646,11 @@ const GameScreen: React.FC<GameScreenProps> = ({ config, onEndGame }) => {
             onHintUsed={() => setUsedHint(true)}
             onExtraAttempt={() => setExtraAttempt(true)}
           />
+          {currentWord.phonemes && (
+            <div className="mt-6">
+              <PhonicsBreakdown phonemes={currentWord.phonemes} />
+            </div>
+          )}
           <div className="flex gap-2 justify-center mb-8">
             {letters.map((letter, idx) => (
               <div
